+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 2023-12-11 - 1.2.3
 - Added an option to config that allow taking outputs from the side of a delivery table. Off by default.
+- Added config option to disable moving Completed Agreement to the output slots.
 - Added several settings to the Delivery Table NBT data for modpack developers. More details on the [Wiki](https://github.com/mortuusars/Wares/wiki/Delivery-Table)
 
 2023-11-10 - 1.2.2